+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod base64_string;
 pub mod da;
 pub mod sequencer;
 pub mod sequencer_block;
